requirements.txt
requests
python-dotenv
Flask
Flask-Cors
PyPDF2
python-docx
fastapi
uvicorn[standard]==0.23.2
pg8000
selenium==4.15.2
webdriver-manager
undetected-chromedriver
pydantic[email]
python-jose[pycryptodome]
passlib[bcrypt]
beautifulsoup4==4.12.3
